+++ index.js
@@ -1,0 +1,24 @@
+/*
+
+	node-utils by SWR audio lab
+	frequently used node-packages and helpers
+
+*/
+
+// import packages
+const ard = require('./packages/ard')
+const helpers = require('./packages/helpers')
+const numbers = require('./packages/numbers')
+const storage = require('./packages/storage-wrapper')
+const strings = require('./packages/strings')
+const undici = require('./packages/undici')
+
+// export packages
+module.exports = {
+	ard,
+	helpers,
+	numbers,
+	storage,
+	strings,
+	undici,
+}