+++ packages/date/index.js
@@ -1,0 +1,21 @@
+// import packages
+const getDateHourMinutes = require('../../utils/date/getDateHourMinutes')
+const getDayMonthYear = require('../../utils/date/getDayMonthYear')
+const getFullRelativeTime = require('../../utils/date/getFullRelativeTime')
+const getHourMinutes = require('../../utils/date/getHourMinutes')
+const getIsoRelativeTime = require('../../utils/date/getIsoRelativeTime')
+const getRelativeTime = require('../../utils/date/getRelativeTime')
+const getYearMonthDay = require('../../utils/date/getYearMonthDay')
+const revYearMonthDay = require('../../utils/date/revYearMonthDay')
+
+// export packages
+module.exports = {
+	getDateHourMinutes,
+	getDayMonthYear,
+	getFullRelativeTime,
+	getHourMinutes,
+	getIsoRelativeTime,
+	getRelativeTime,
+	getYearMonthDay,
+	revYearMonthDay,
+}