+++ packages/helpers/index.js
@@ -1,7 +1,11 @@
 // import packages
+const arrayToObjectCount = require('../../utils/helpers/arrayToObjectCount')
+const getJsonKeys = require('../../utils/helpers/getJsonKeys')
 const sleep = require('../../utils/helpers/sleep')
 
 // export packages
 module.exports = {
+	arrayToObjectCount,
+	getJsonKeys,
 	sleep,
 }