+++ packages/strings/index.js
@@ -4,6 +4,7 @@ const getObjectLength = require('../../utils/strings/getObjectLength')
 const isArray = require('../../utils/strings/isArray')
 const isEmptyArray = require('../../utils/strings/isEmptyArray')
 const isEmptyObject = require('../../utils/strings/isEmptyObject')
+const isEmptyString = require('../../utils/strings/isEmptyString')
 const isIncluded = require('../../utils/strings/isIncluded')
 const isNull = require('../../utils/strings/isNull')
 const isObject = require('../../utils/strings/isObject')
@@ -22,6 +23,7 @@ module.exports = {
 	isArray,
 	isEmptyArray,
 	isEmptyObject,
+	isEmptyString,
 	isIncluded,
 	isNull,
 	isObject,