+++ utils/date/getHourMinutes.js
@@ -1,0 +1,4 @@
+const { DateTime } = require('luxon')
+
+// get hours and minutes (returns '12:34')
+module.exports = (date) => DateTime.fromISO(date).toFormat('HH:mm')