+++ utils/numbers/normalize.js
@@ -1,0 +1,2 @@
+// normalize value by division of a maximum value
+module.exports = (value, maximumValue) => value / maximumValue